feat(rotas): renomear /dashboard → /today, /stats → /progress
- Criar today.vue e progress.vue com conteúdo migrado
- dashboard.vue e stats.vue viram redirects 301
- Atualizar todas as referências internas (/dashboard → /today, /decks → /topics)
- Atualizar rota pós-login para /today

Refs: #40
